fastlio2: fov_degree is an int in the generated YAML
FAST-LIO's yaml-cpp reads mapping/fov_degree as int; emitting 360.0 threw
TypedBadConversion<int> and aborted the binary. Caught by an end-to-end VM
replay (acc_cov=1.0 -> bounded 4.8 m over 25 s, rrd written).

+++ dimos/hardware/sensors/lidar/fastlio2/module.py
@@ -144,7 +144,7 @@ class FastLio2Config(NativeModuleConfig):
     b_gyr_cov: float = 0.0001
     filter_size_surf: float = 0.1  # IESKF scan voxel; does not affect divergence
     filter_size_map: float = 0.1
-    fov_degree: float = 360.0
+    fov_degree: int = 360  # FAST-LIO reads this as an int
     det_range: float = 100.0
     extrinsic_est_en: bool = False  # online IMU-LiDAR extrinsic estimation
     extrinsic_t: list[float] = Field(default_factory=lambda: [-0.011, -0.02329, 0.04412])
